refactor: split bm25/idle_evict test modules to satisfy the 500-line file cap
CI's line-cap gate flagged two files pushed over the ratcheted cap by this
change's new tests:
- trusty-common/src/bm25.rs (503 SLOC, production-classified since it mixes
  Bm25Index + its own inline test module) -- split the inline `mod tests`
  block out to bm25/tests.rs (matches the exact "tests.rs" test-file
  pattern the line-cap script recognizes).
- core/indexer/tests_idle_evict.rs (659 SLOC after the new #3683/#3684
  tests -- this file's "tests_*" prefix doesn't match the line-cap script's
  test-file suffix patterns, so it's production-capped at 500 despite being
  pure test content) -- moved the three new rehydrate tests
  (detached_rehydrate_survives_caller_cancellation,
  rehydrate_dedupes_concurrent_callers_onto_one_scan,
  rehydrate_is_deterministic_across_repeated_cycles_over_cap) to a new
  sibling file core/indexer/rehydrate_tests.rs, whose "_tests.rs" suffix
  does match, giving it the 1500-SLOC test budget.

No behavior change -- pure file split, mirroring the existing
search/lanes_tests.rs convention of duplicating small test-only helpers
(make_indexer_with_corpus) per file rather than sharing them cross-file.
